Move components from root to elements
