+++ web/components/elements/Button.tsx
@@ -1,7 +1,7 @@
-import { h } from "preact";
+import { FunctionComponent, h } from "preact";
 import style from "./Button.module.scss";
 
-export function Button(props: { [key: string]: unknown, intent?: string}) {
+export const Button: FunctionComponent = (props: { [key: string]: unknown, intent?: string}) => {
     let className = style.button;
     if (props.intent === "remove") {
         className += ` ${  style.remove}`;
