+++ web/components/elements/ButtonSet.tsx
@@ -1,9 +1,8 @@
 import { FunctionComponent, h } from "preact";
 import style from "./ButtonSet.module.scss";
 
-const ButtonSet: FunctionComponent = (props) => {
+export const ButtonSet: FunctionComponent = (props) => {
     return <div className={style.buttonSet}>
         {props.children}
     </div>;
 }
-export default ButtonSet;
