+++ web/components/elements/InputField.tsx
@@ -7,11 +7,9 @@ interface Props {
     noPadding: boolean;
 }
 
-const InputField: FunctionComponent<Props> = ({ children, visible = true, label, noPadding }) => {
+export const InputField: FunctionComponent<Props> = ({ children, visible = true, label, noPadding }) => {
     return visible && <div className={style.inputField}>
         {label && <label className={noPadding ? style.nopad : ""}>{label}</label>}
         {children}
     </div>;
 };
-
-export default InputField;
